changed the divide case removing casting

+++ src/main/java/ie/atu/week2/RequestController.java
@@ -44,7 +44,7 @@ public class RequestController
                 }
                 else{
                     operationMode="divide";
-                    ans= (float)num1/num2;
+                    ans= num1/num2;
                 }
                 break;
             case "multiply":
